+++ src/cljs/cxengage_javascript_sdk/modules/voice.cljs
@@ -113,9 +113,10 @@
        (p/publish {:topic (:topic interrupt-params)
                    :error (e/invalid-args-error (s/explain-data (:validation interrupt-params) client-params))
                    :callback callback})
-       (do #_(when (or (= type :transfer-to-resource)
-                       (= type :transfer-to-queue)
-                       (= type :transfer-to-extension))
+       (do (when (and (= transfer-type "warm-transfer")
+                      (or (= type :transfer-to-resource)
+                          (= type :transfer-to-queue)
+                          (= type :transfer-to-extension)))
                (send-interrupt module :hold {:interaction-id interaction-id}))
            (iu/send-interrupt* module (assoc interrupt-params
                                              :interaction-id interaction-id